Deduplicate some command endpoint models.
