+++ robot-server/robot_server/runs/router/commands_router.py
@@ -1,13 +1,11 @@
 """Router for /runs commands endpoints."""
 import textwrap
-from datetime import datetime
 from typing import Optional, Union
 from typing_extensions import Final, Literal
 
 from anyio import move_on_after
 from fastapi import APIRouter, Depends, Query, status
 
-from pydantic import BaseModel, Field
 
 from opentrons.protocol_engine import (
     ProtocolEngine,
@@ -17,7 +15,6 @@ from opentrons.protocol_engine import (
 
 from robot_server.errors.error_responses import ErrorDetails, ErrorBody
 from robot_server.service.json_api import (
-    RequestModel,
     SimpleBody,
     MultiBody,
     MultiBodyMeta,
@@ -26,6 +23,12 @@ from robot_server.service.json_api import (
 )
 from robot_server.robot.control.dependencies import require_estop_in_good_state
 
+from ..command_models import (
+    RequestModelWithCommandCreate,
+    CommandCollectionLinks,
+    CommandLink,
+    CommandLinkMeta,
+)
 from ..run_models import RunCommandSummary
 from ..run_data_manager import RunDataManager, PreSerializedCommandsNotAvailableError
 from ..engine_store import EngineStore
@@ -38,17 +41,6 @@ from .base_router import RunNotFound, RunStopped
 _DEFAULT_COMMAND_LIST_LENGTH: Final = 20
 
 commands_router = APIRouter()
-
-
-class RequestModelWithCommandCreate(RequestModel[pe_commands.CommandCreate]):
-    """Equivalent to RequestModel[CommandCreate].
-
-    This works around a Pydantic v<2 bug where RequestModel[CommandCreate]
-    doesn't parse using the CommandCreate union discriminator.
-    https://github.com/pydantic/pydantic/issues/3782
-    """
-
-    data: pe_commands.CommandCreate
 
 
 class CommandNotFound(ErrorDetails):
@@ -81,35 +73,6 @@ class PreSerializedCommandsNotAvailable(ErrorDetails):
     title: str = "Pre-Serialized commands not available."
     detail: str = (
         "Pre-serialized commands are only available once a run has finished running."
-    )
-
-
-class CommandLinkMeta(BaseModel):
-    """Metadata about a command resource referenced in `links`."""
-
-    runId: str = Field(..., description="The ID of the command's run.")
-    commandId: str = Field(..., description="The ID of the command.")
-    index: int = Field(..., description="Index of the command in the overall list.")
-    key: str = Field(..., description="Value of the current command's `key` field.")
-    createdAt: datetime = Field(
-        ...,
-        description="When the current command was created.",
-    )
-
-
-class CommandLink(BaseModel):
-    """A link to a command resource."""
-
-    href: str = Field(..., description="The path to a command")
-    meta: CommandLinkMeta = Field(..., description="Information about the command.")
-
-
-class CommandCollectionLinks(BaseModel):
-    """Links returned along with a collection of commands."""
-
-    current: Optional[CommandLink] = Field(
-        None,
-        description="Path to the currently running or next queued command.",
     )
