+++ robot-server/tests/runs/router/test_commands_router.py
@@ -16,15 +16,17 @@ from opentrons.protocol_engine import (
 from robot_server.errors.error_responses import ApiError
 from robot_server.service.json_api import MultiBodyMeta
 
+from robot_server.runs.command_models import (
+    RequestModelWithCommandCreate,
+    CommandCollectionLinks,
+    CommandLink,
+    CommandLinkMeta,
+)
 from robot_server.runs.run_store import RunStore, CommandNotFoundError
 from robot_server.runs.engine_store import EngineStore
 from robot_server.runs.run_data_manager import RunDataManager
 from robot_server.runs.run_models import RunCommandSummary, RunNotFoundError
 from robot_server.runs.router.commands_router import (
-    CommandCollectionLinks,
-    CommandLink,
-    CommandLinkMeta,
-    RequestModelWithCommandCreate,
     create_run_command,
     get_run_command,
     get_run_commands,
